requirements.txt
keras==2.6.0
tensorflow-cpu==2.6.0
Pillow==8.2.0
matplotlib
numpy==1.19.5
scikit-learn==0.24.2
pandas
flask

